Test code cleanup
* Remove deprecated TensorGen (no longer used)
* Change TestUtil to be package object in dimwit % Test. This makes
  Global Test utilities Predef in all tests, meaning no import is
required.
