+++ core/src/test/scala/dimwit/tensor/TensorOpsBinarySuite.scala
@@ -2,7 +2,6 @@ package dimwit.tensor
 
 import dimwit.*
 import dimwit.Conversions.given
-import TestUtil.*
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.funspec.AnyFunSpec
